mapping Patch-bugs

+++ src/main/java/dev/repository/MissionRepo.java
@@ -9,7 +9,7 @@ import org.springframework.data.jpa.repository.Query;
 import dev.domain.Collegue;
 import dev.domain.Mission;
 
-public interface MissionRepo extends JpaRepository<Mission, Long> {
+public interface MissionRepo extends JpaRepository<Mission, Integer> {
 
 	List<Mission> findByCollegue(Collegue collegue);
